LPS-57519 Move Action related API classes
