+++ modules/apps/mobile-device-rules/mobile-device-rules-api/src/com/liferay/mobile/device/rules/action/ActionHandlerManager.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.kernel.mobile.device.rulegroup.action;
+package com.liferay.mobile.device.rules.action;
 
+import com.liferay.mobile.device.rules.model.MDRAction;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portlet.mobiledevicerules.model.MDRAction;
 
 import java.util.Collection;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -25,15 +26,21 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * @author Edward Han
  */
-public interface ActionHandler {
+public interface ActionHandlerManager {
 
-	public void applyAction(
-			MDRAction mdrAction, HttpServletRequest request,
+	public void applyActions(
+			List<MDRAction> mdrActions, HttpServletRequest request,
 			HttpServletResponse response)
 		throws PortalException;
 
-	public Collection<String> getPropertyNames();
+	public ActionHandler getActionHandler(String actionType);
 
-	public String getType();
+	public Collection<ActionHandler> getActionHandlers();
+
+	public Collection<String> getActionHandlerTypes();
+
+	public void registerActionHandler(ActionHandler actionHandler);
+
+	public ActionHandler unregisterActionHandler(String actionType);
 
 }
