--- portal-service/src/com/liferay/portal/kernel/mobile/device/rulegroup/action/packageinfo
@@ -1,1 +1,0 @@
-version 6.2.0
